gui/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
lxml==5.2.1
openpyxl==3.1.2
packaging==24.0
python-docx==1.1.0
typing_extensions==4.11.0